updated test cases for feeRate
